% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_summary.R
\name{daily_summary}
\alias{daily_summary}
\title{Summarise daily mean FoI by scenario}
\usage{
daily_summary(x)
}
\arguments{
\item{x}{data frame with columns: integrated_FoI, days_infectious_community, scenario}
}
\value{
tibble with scenario, day, mean_integrated_FoI
}
\description{
Summarise daily mean FoI by scenario
}
