% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarrisk-package.R
\docType{package}
\name{quarrisk-package}
\alias{quarrisk}
\alias{quarrisk-package}
\title{quarrisk: Quarantine Risk Explorer}
\description{
The \code{quarrisk} package provides tools and an interactive Shiny dashboard
to explore simulated quarantine breach and infection risk under different
Delta variant scenarios. It combines data, visualization, and interactivity
to help understand how quarantine measures, vaccination, and compliance
affect outbreak likelihood.
}
\author{
\strong{Maintainer}: Siddhi Jadhav \email{sjad0009@student.monash.edu}

}
\keyword{internal}
